replace SBCS by MBCS
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/484



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@3297 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/common/ucnvmbcs.h
@@ -158,7 +158,7 @@ _MBCSSingleSimpleGetNextUChar(UConverterSharedData *sharedData,
  * returns fallback values.
  */
 #define _MBCS_SINGLE_SIMPLE_GET_NEXT_BMP(sharedData, b) \
-    (UChar)(((sharedData)->table->mbcs.stateTable[0][b])>>7)
+    (UChar)(((sharedData)->table->mbcs.stateTable[0][(uint8_t)(b)])>>7)
 
 /**
  * This is an internal function that allows other converter implementations
